+++ Management/SoundRanks.cs
@@ -1,0 +1,13 @@
+namespace UltraRankSounds.Management;
+
+public class SoundRanks
+{
+    public static readonly int DESTRUCTIVE = 0;
+    public static readonly int CHAOTIC = 1;
+    public static readonly int BRUTAL = 2;
+    public static readonly int ANARCHIC = 3;
+    public static readonly int SUPREME = 4;
+    public static readonly int SSADISTIC = 5;
+    public static readonly int SSSHITSTORM = 6;
+    public static readonly int ULTRAKILL = 7;
+}