+++ Patches/StyleHUDPatch.cs
@@ -2,6 +2,7 @@ using HarmonyLib;
 using UnityEngine;
 using UltraRankSounds.Components;
 using BepInEx;
+using System.Collections.Generic;
 
 namespace UltraRankSounds.Patches
 {
@@ -9,14 +10,23 @@ namespace UltraRankSounds.Patches
     [HarmonyPatch(typeof(StyleHUD))]
     public class StyleHUDPatch
     {
-        private static CustomSoundPlayer customSoundPlayer = null;
+        private static CustomSoundPlayer styleRankSoundPlayer = null;
+        private static QueuedCustomSoundPlayer styleBonusSoundPlayer = null;
         private static int lastRankIndex = 0;
 
         [HarmonyPostfix]
         [HarmonyPatch(typeof(StyleHUD), "Start")]
         public static void StyleHUD_Start_Postfix(GameObject ___styleHud, StyleHUD __instance)
         {
-            customSoundPlayer = ___styleHud.AddComponent<CustomSoundPlayer>();
+            styleRankSoundPlayer = ___styleHud.AddComponent<CustomSoundPlayer>();
+            styleBonusSoundPlayer = ___styleHud.AddComponent<QueuedCustomSoundPlayer>();
+        }
+
+        [HarmonyPostfix]
+        [HarmonyPatch(typeof(StyleHUD), "Awake")]
+        public static void StyleHUD_Awake_Postfix(Dictionary<string,string> ___idNameDict)
+        {
+            UltraRankSounds.RegisterStylePointBonuses(___idNameDict);
         }
 
         [HarmonyPostfix]
@@ -27,6 +37,22 @@ namespace UltraRankSounds.Patches
                 return;
 
             PlaySoundForRank(0, "RankD");
+        }
+
+        [HarmonyPostfix]
+        [HarmonyPatch(typeof(StyleHUD), "AddPoints")]
+        public static void StyleHUD_AddPoints_Postfix(string pointID, StyleHUD __instance)
+        {
+            if (!UltraRankSounds.EnableSounds.value || !UltraRankSounds.EnableStyleBonusSounds.value)
+                return;
+
+            string localizedName = __instance.GetLocalizedName(pointID);
+
+            if (string.IsNullOrWhiteSpace(localizedName) || string.IsNullOrWhiteSpace(pointID))
+                return;
+
+            UltraRankSounds.Log($"got style bonus '{localizedName}' with id '{pointID}'");
+            PlaySoundForBonus(pointID);
         }
 
         [HarmonyPostfix]
@@ -58,14 +84,30 @@ namespace UltraRankSounds.Patches
 
         private static void PlaySoundForRank(int index, string name, bool ascended = true)
         {
-            if (!UltraRankSounds.EnableSounds.value || customSoundPlayer == null)
+            if (!UltraRankSounds.EnableSounds.value || styleRankSoundPlayer == null)
                 return;
 
             string soundFile = GetSoundPathForRankIndex(index, ascended);
             string status = ascended ? "ascended" : "descended";
 
             UltraRankSounds.Log($"{status} style rank to '{name}' [index: {index} oldindex: {lastRankIndex}] | playing sound '{soundFile}'");
-            customSoundPlayer.PlaySound(soundFile);
+            styleRankSoundPlayer.PlaySound(soundFile);
+        }
+
+        private static void PlaySoundForBonus(string id)
+        {
+            if (!UltraRankSounds.EnableSounds.value || styleBonusSoundPlayer == null)
+                return;
+
+            string soundFile = Management.SoundConfig.GetPointBonusSound(id);
+            if (string.IsNullOrEmpty(soundFile))
+            {
+                UltraRankSounds.Log($"Unable to find rank sound for point bonus '{id}'", true);
+                return;
+            }
+
+            UltraRankSounds.Log($"{id} style bonus registered | playing sound '{soundFile}'");
+            styleBonusSoundPlayer.QueueSound(soundFile);
         }
 
         private static string GetSoundPathForRankIndex(int rank, bool ascended)
@@ -73,11 +115,10 @@ namespace UltraRankSounds.Patches
             int r = rank;
 
             if (ascended)
-                return SoundsConfig.GetAscensionRankSoundName(r);
+                return Management.SoundConfig.GetAscensionRankSoundName(r);
 
-            return SoundsConfig.GetDescensionRankSoundName(r);
+            return Management.SoundConfig.GetDescensionRankSoundName(r);
         }
 
     }
-
 }